[OPT]update to zy_compass_3.4 :
    1. support per-layer configuration (some fields) in cfg file
    2. dataset plugin can get graph's output tensors of previous forward
    3. add operators: atan, col2im, BatchToSpaceND, AdaptivePool, FilterBox
    4. simulate AIFF's ahead shift for 16bit operators
    5. refine operators: DecodeBox, DetectionOutput, LayerNorm, MatMul, MaxPoolingWithArgMax, Negative, Pad, Proposal, PyramidROIAlign, ROIAlgin, BasicLSTM, GRU
    6. bugs fix

+++ AIPUBuilder/Optimizer/ops/cumulate.py
@@ -19,8 +19,8 @@ def convert_less_mbit(value, threshold_min, threshold_max):
     while (data.max() > threshold_max) or (data.min() < threshold_min):
         max_mask = data > threshold_max
         min_mask = data < threshold_min
-        data[max_mask] = data[max_mask] >> 1
-        data[min_mask] = data[min_mask] >> 1
+        data[max_mask] = (data[max_mask].long() >> 1).to(data.dtype)
+        data[min_mask] = (data[min_mask].long() >> 1).to(data.dtype)
         left_shift[max_mask] += 1
         left_shift[min_mask] += 1
     if data.dim() < 1:
